+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+
+- enabling check BaseInput, BaseTextarea, BaseSelect after connection to page
+
+### Changed
+
+- 'display:none' with 'padding:0' in error message of BaseInput, BaseTextarea and BaseSelect
+
 ## [0.6.0]
 
 ### Added