+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - 'display:none' with 'padding:0' in error message of BaseInput, BaseTextarea and BaseSelect
+- setting boolean attributes (required, etc.)
 
 ### Fixed